refactor: rewrite controllers for class
